reader-viaccess.c :  change a log message for pin code 	oscam.c : fix a compiler warning (dependent on gcc version). 	reader-seca.c : inplement the same pin code processing as in reader-viaccess.c
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@3827 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ oscam.c
@@ -273,7 +273,7 @@ struct s_client * get_client_by_tid(unsigned long tid) //FIXME untested!! no lon
 {
   struct s_client *cl; 
   for (cl=first_client; cl ; cl=cl->next)
-    if (cl->thread==tid)
+    if ((unsigned int)(cl->thread)==tid)
       return cl;
   return NULL;
 }
